backend/requirements.txt
appdirs==1.4.4
attrs==22.1.0
cattrs==22.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
CoinMarketCapAPI==0.5
exceptiongroup==1.0.0rc9
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
greenlet==1.1.3.post0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
load-dotenv==0.1.0
MarkupSafe==2.1.1
psycopg2-binary==2.9.4
python-coinmarketcap==0.3
python-dotenv==0.21.0
ratelimit==2.2.1
requests==2.28.1
requests-cache==0.9.6
six==1.16.0
SQLAlchemy==1.4.42
url-normalize==1.4.3
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.10.0